Last merge

+++ Fase3/BreadSpread/BreadSpread/Controllers/ManutencaoController.cs
@@ -4,10 +4,11 @@ using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using BreadSpread.Models;
 using BreadSpread.Controllers;
+using System.Net.Mail;
+using BreadSpread.Encriptacao;
 
 namespace BreadSpread.Controllers
 {
@@ -17,7 +18,193 @@ namespace BreadSpread.Controllers
 
         public ActionResult AdminIndex()
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
             return View("~/Views/Admin/Index.cshtml");
+        }
+
+        public ActionResult Perfil()
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
+            var User_In_Session = User.Identity.Name;
+
+            var cliente = db.Clientes.Where(c => c.email.Equals(User_In_Session)).ToList();
+
+            if (cliente == null)
+            {
+                return HttpNotFound();
+            }
+            return View(cliente[0]);
+        }
+
+        private void Send(Encomenda e)
+        {
+            using (MailMessage mm = new MailMessage("breadspread365@gmail.com", User.Identity.Name))
+            {
+                mm.Subject = "Confirmação de pagamento";
+                mm.Body = "O presente email confirma que a encomenda número " + 
+                           e.idEnc + " foi paga no dia " + e.dataPag + ".";
+                mm.IsBodyHtml = false;
+                using (SmtpClient smtp = new SmtpClient())
+                {
+                    smtp.Host = "smtp.gmail.com";
+                    smtp.EnableSsl = true;
+                    NetworkCredential NetworkCred = new NetworkCredential("breadspread365@gmail.com", "vamosfazerli4");
+                    smtp.UseDefaultCredentials = true;
+                    smtp.Credentials = NetworkCred;
+                    smtp.Port = 587;
+                    smtp.Send(mm);
+                }
+            }
+        }
+
+        public ActionResult Paga(int id)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
+            Encomenda e = db.Encomendas.Find(id);
+            e.dataPag = DateTime.Now;
+            e.modoPag = "online";
+            e.estado = "confirmada";
+            if (ModelState.IsValid)
+            {
+                db.Entry(e).State = EntityState.Modified;
+                db.SaveChanges();
+                Send(e);
+            }
+
+			return Redirect(Request.UrlReferrer.ToString());
+		}
+
+        public ActionResult DesativaCliente(int id)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
+            Cliente c = db.Clientes.Find(id);
+            c.estadoConta = "desativado";
+            db.Entry(c).State = EntityState.Modified;
+            db.SaveChanges();
+
+            return RedirectToAction("Logout", "Autenticacao");
+        }
+
+        public ActionResult PagaEncomendas(int id)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
+            return View(db.Encomendas.Where(enc => enc.idCli.Equals(id) && 
+                                            (enc.estado.Equals("pendente") || enc.estado.Equals("confirmada")) && 
+                                            enc.dataPag == null)
+                                     .ToList());
+        }
+
+        public ActionResult EditPerfil(int? id)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Cliente cliente = db.Clientes.Find(id);
+            if (cliente == null)
+            {
+                return HttpNotFound();
+            }
+            return View(cliente);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditPerfil([Bind(Include = "idCli,nome,email,sexo,dataNasc,rua,freguesia,cidade,codPostal,numPorta,contacto,NIF, estadoConta,password")] Cliente cliente)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            //cliente.estadoConta = "ativo";
+            Cliente c = db.Clientes.Find(cliente.idCli);
+            if (!cliente.password.Equals(c.password))
+            {
+                cliente.password = MyHelpers.HashPassword(cliente.password);
+            }
+            db.Entry(c).State = EntityState.Detached;
+            //DEIXEI ASSIM PORQUE COM O MODEL STATE NÃO DÁ (COMO ESTÁ INSERE CORRETAMENTE TUDO)
+            if (ModelState.IsValid)
+            {
+                db.Entry(cliente).State = EntityState.Modified;
+                db.SaveChanges();
+            }
+            return RedirectToAction("Perfil");
+        }
+
+        public ActionResult Ocasionais()
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            var encomendas = db.Encomendas.Where(e => e.estado.Equals("pendente")).ToList();
+            /*var encomendas = (from e in db.Encomendas
+                            where e.estado == "pendente"
+                            select e);*/
+
+            return View(encomendas);
+        }
+
+        public ActionResult AceitarEncomenda(int? idEnc)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            if (idEnc == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Encomenda encomenda = db.Encomendas.Find(idEnc);
+            if (encomenda == null)
+            {
+                return HttpNotFound();
+            }
+
+            encomenda.estado = "confirmada";
+
+            if (ModelState.IsValid)
+            {
+                db.Entry(encomenda).State = EntityState.Modified;
+                db.SaveChanges();
+            }
+            return RedirectToAction("Ocasionais");
+        }
+
+        public ActionResult RecusarEncomenda(int? idEnc)
+        {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            if (idEnc == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Encomenda encomenda = db.Encomendas.Find(idEnc);
+            if (encomenda == null)
+            {
+                return HttpNotFound();
+            }
+
+            //remoçºao da encomenda em si
+            db.Encomendas.Remove(encomenda);
+
+            //remoçºao de todas as entradas associadas à encomenda a remover
+            var enc_rem = db.Encomenda_Produto.Where(e => e.idEnc == idEnc).ToList();
+            foreach (var m in enc_rem)
+                db.Encomenda_Produto.Remove(m);
+
+            db.SaveChanges();
+
+            return RedirectToAction("Ocasionais");
         }
 
         //==============================
@@ -27,12 +214,16 @@ namespace BreadSpread.Controllers
         // GET: Manutencao
         public ActionResult IndexProduto()
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             return View(db.Produtoes.ToList());
         }
 
         // GET: Manutencao/Details/5
         public ActionResult DetailsProduto(int? id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -48,6 +239,8 @@ namespace BreadSpread.Controllers
         // GET: Manutencao/Create
         public ActionResult CreateProduto()
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             return View();
         }
 
@@ -58,6 +251,8 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult CreateProduto([Bind(Include = "designacao,ingredientes,infoNutricional,preco,imagem,peso")] Produto produto)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (ModelState.IsValid)
             {
                 db.Produtoes.Add(produto);
@@ -71,6 +266,8 @@ namespace BreadSpread.Controllers
         // GET: Manutencao/Edit/5
         public ActionResult EditProduto(int? id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -90,6 +287,8 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult EditProduto([Bind(Include = "idProd,designacao,ingredientes,infoNutricional,preco,imagem,peso")] Produto produto)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (ModelState.IsValid)
             {
                 db.Entry(produto).State = EntityState.Modified;
@@ -102,6 +301,8 @@ namespace BreadSpread.Controllers
         // GET: Manutencao/Delete/5
         public ActionResult DeleteProduto(int? id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -119,6 +320,8 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmedProduto(int id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             Produto produto = db.Produtoes.Find(id);
             db.Produtoes.Remove(produto);
             db.SaveChanges();
@@ -133,12 +336,18 @@ namespace BreadSpread.Controllers
         // GET: Funcionarios
         public ActionResult IndexFuncionario()
         {
-            return View(db.Funcionarios.ToList());
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            List<Funcionario> funcs = db.Funcionarios.Where(f => f.estadoConta.Equals("ativo")).ToList();
+
+            return View(funcs);
         }
 
         // GET: Funcionarios/Details/5
         public ActionResult DetailsFuncionario(string id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -154,6 +363,8 @@ namespace BreadSpread.Controllers
         // GET: Funcionarios/Create
         public ActionResult CreateFuncionario()
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             return View();
         }
 
@@ -164,6 +375,9 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult CreateFuncionario([Bind(Include = "idFunc,nome,dataNasc,contacto,rua,numPorta,codPostal,cidade,password,freguesia,estadoConta,distribuicao")] Funcionario funcionario)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+            funcionario.password = MyHelpers.HashPassword(funcionario.password);
             if (ModelState.IsValid)
             {
                 db.Funcionarios.Add(funcionario);
@@ -177,6 +391,8 @@ namespace BreadSpread.Controllers
         // GET: Funcionarios/Edit/5
         public ActionResult EditFuncionario(string id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -196,6 +412,8 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult EditFuncionario([Bind(Include = "idFunc,nome,dataNasc,contacto,rua,numPorta,codPostal,cidade,password,freguesia,estadoConta,distribuicao")] Funcionario funcionario)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             if (ModelState.IsValid)
             {
                 db.Entry(funcionario).State = EntityState.Modified;
@@ -208,6 +426,9 @@ namespace BreadSpread.Controllers
         // GET: Funcionarios/Delete/5
         public ActionResult DeleteFuncionario(string id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -225,8 +446,11 @@ namespace BreadSpread.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmedFuncionario(string id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return RedirectToAction("Index", "Autenticacao");
             Funcionario funcionario = db.Funcionarios.Find(id);
-            db.Funcionarios.Remove(funcionario);
+            funcionario.estadoConta = "desativado";
+            db.Entry(funcionario).State = EntityState.Modified;
             db.SaveChanges();
             return RedirectToAction("IndexFuncionario");
         }
